refactor(core) astubbs#228: the two allocators share one ledger-backed body, and differ only in the division
The PR's duplicate-code report named the clones this branch introduced: the
partition-share allocator and the in-process one carried the same registry,
the same ledger behind the same monitor, and the same bodies for the quantum
pull, the soft debit, the lease read, the conservation snapshot and the
resource-wide reads - "find the contract, index the quantum, act on the ledger
under the monitor" written twice. That body now lives once in a package-private
abstract base, and each allocator supplies only what makes it itself: what a
member's share of a quantum IS (the rotation over held ordinals under the
effective snapshot, or the rotation over the issued membership, or not a
participant at all), what burst budget a debit is judged against, and two
hooks under the same monitor for the in-process rung's lease renewal and
membership pruning. Membership, the member-side next-credit projection and the
local rate stay with each subclass, because those are the division.

No behaviour moved: both allocators' unit suites, both Lincheck harnesses, the
rebalance-callback tests and the two-JVM lane are green unchanged. The one
shape Error Prone refused - a subclass calling super under the inherited guard
- became a final settle plus an overridable hook, which is also the clearer
shape. The rebalance test that reaches the monitor by reflection now walks the
hierarchy to find it.
